Changes for arm64 and Clang redefined-types

+++ Signal/src/audio/incall_audio/RemoteIOBufferListWrapper.m
@@ -8,7 +8,7 @@
     AudioBufferList* audioBufferList = malloc(sizeof(AudioBufferList));
     audioBufferList->mNumberBuffers = 1;
     audioBufferList->mBuffers[0].mNumberChannels = 1;
-    audioBufferList->mBuffers[0].mDataByteSize = bufferSize;
+    audioBufferList->mBuffers[0].mDataByteSize = (UInt32)bufferSize;
     audioBufferList->mBuffers[0].mData = malloc(bufferSize);
     
     RemoteIOBufferListWrapper* w = [RemoteIOBufferListWrapper new];
